move executor to mlas directory
