+++ src/plugins/intel_cpu/src/nodes/executors/transpose_list.hpp
@@ -13,7 +13,7 @@
 
 #include "common/ref_opt_transpose.hpp"
 #include "common/ref_transpose.hpp"
-#include "common/mlas_transpose.hpp"
+#include "mlas/mlas_transpose.hpp"
 #include "x64/jit_transpose.hpp"
 
 #include "onednn/iml_type_mapper.h"
